# api/api.yaml
openapi: 3.0.1  
info:
  version: 1.0.0
  title: Hacker News
  description: "This is the Open API documentation for the REST API of our beloved application **Hacker News** deployed at <https://hacker-news-fib.herokuapp.com/>. <br>All operations are executable. Only one operation requires authentication: `deleteTweet`. In this case, you must **Authorize** your request by providing the api_key vaule you got when you created the tweet."
servers:
  - url: 'http://127.0.0.1:3000/api/v1/'
paths:
...
  /reply/{id}/vote:
      post:
        tags:
          - votes
          - replies
        summary: Vote a Reply
        description: Vote a Reply
        operationId: voteReply
        parameters:
        - name: id
          in: path
          description: ID of Reply to return
          required: true
          schema:
            type: integer
            format: int64
        responses:
          201:
            description: Vote done
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Reply'
          401:
            description: 'Error: Unauthorized'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error401'
          403:
            description: 'Error: Forbidden'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error403'
          404:
            description: 'Error: Not Found'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error404Reply'
          409:
            description: 'Error: Conflict'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error409Vote'
        security:
          - ApiKeyAuth: []
      delete:
        tags:
          - replies
          - votes
        summary: Unvote a Reply
        description: Unvote a Reply
        operationId: unvoteReply
        parameters:
        - name: id
          in: path
          description: ID of Reply to return
          required: true
          schema:
            type: integer
            format: int64
        responses:
          204:
            description: Unvote done
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Reply'
          401:
            description: 'Error: Unauthorized'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error401'
          403:
            description: 'Error: Forbidden'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error403'
          404:
            description: 'Error: Not Found'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error404Post'
          409:
            description: 'Error: Conflict'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  errorExample:
                    $ref: '#/components/examples/error409Unvote'
        security:
          - ApiKeyAuth: []

### COMPONENTS
components:
  schemas:
...
    Reply:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        text:
          type: string
          minLength: 4
          example: 'Reply Text'
        votes:
          type: integer
          format: int64
          example: 5
        userId:
          type: integer
          format: int64
          example: 3
        parentId:
          type: integer
          format: int64
          example: 1
        parentType:
          type: string
          minLength: 4
          example: 'Comment'
        created_at:
          type: string
          format: date-time
...
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 403, 404, 409]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Forbidden", "Not Found", "Conflict"]
        message: 
          type: string
  examples:
...
    error401:
      value:
        status: 401
        error: "Unauthorized"
        message: "You provided no api key (X-API-KEY Header)"
    error403:
      value:
        status: 403
        error: "Forbidden"
        message: "Your api key (X-API-KEY Header) is not valid"
...
    error404Reply:
      value:
        status: 404
        error: "Not Found"
        message: "No Reply with that ID"
    error404Post:
      value:
        status: 404
        error: "Not Found"
        message: "No Post with that ID"
...
    error409Vote:
      value:
        status: 409
        error: "Conflict"
        message: "Vote to the same Contribution already created"
    error409Unvote:
      value:
        status: 409
        error: "Conflict"
        message: "No Vote to the Contribution exists"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY
      in: header